pyproject.toml
[tool.poetry]
name = "ampter"
version = "0.1.0"
description = "Audio Mastering Painter"
authors = ["echometerain <echometer@disroot.org>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
pedalboard = "^0.7.4"
customtkinter = "^5.2.0"
librosa = "^0.10.0.post2"
Pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
